all good, ready for extraction unit testing
