+++ .gitignore
@@ -58,3 +58,5 @@ src/extraction/extract
 quickest_path/data/val_de_loire_roads.csv
 quickest_path/data/extractedDataset.bin
 src/data/extractedDataset.bin
+data/USA-roads.csv
+data/extractedDataset.bin
